packages/pdfsys-layout-analyser/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdfsys-layout-analyser"
version = "0.0.1"
description = "Layout model runner: emits LayoutDocument to cache, runs once per PDF on the needs-ocr branch."
requires-python = ">=3.11"
dependencies = [
    "pdfsys-core",
    "pymupdf>=1.25",
    "doclayout-yolo>=0.0.2",
    "numpy>=1.24",
    "Pillow>=10.0",
    "huggingface-hub>=0.20",
]

[tool.uv.sources]
pdfsys-core = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/pdfsys_layout_analyser"]
